# Performance Profiler: Update the thumbnail container to align to the right